TST: collect unary tests
